+++ src/main/java/com/kingpixel/cobblebosses/events/CaptureEvents.java
@@ -2,6 +2,8 @@ package com.kingpixel.cobblebosses.events;
 
 import com.cobblemon.mod.common.api.Priority;
 import com.cobblemon.mod.common.api.events.CobblemonEvents;
+import com.cobblemon.mod.common.api.pokemon.PokemonProperties;
+import com.cobblemon.mod.common.entity.pokemon.PokemonEntity;
 import com.cobblemon.mod.common.pokemon.Pokemon;
 import com.kingpixel.cobblebosses.CobbleBosses;
 import com.kingpixel.cobblebosses.model.Boss;
@@ -13,13 +15,19 @@ import net.minecraft.server.network.ServerPlayerEntity;
  */
 public class CaptureEvents {
   public static void register() {
+    CobblemonEvents.POKE_BALL_CAPTURE_CALCULATED.subscribe(Priority.LOWEST, evt -> {
+      PokemonEntity pokemonEntity = evt.getPokemonEntity();
+      Pokemon pokemon = pokemonEntity.getPokemon();
+      Boss boss = CobbleBosses.bossesConfig.getBoss(pokemon);
+      if (boss != null && evt.getCaptureResult().isSuccessfulCapture())
+        PokemonProperties.Companion.parse(boss.getProperties()).apply(pokemon);
+    });
     CobblemonEvents.POKEMON_CAPTURED.subscribe(Priority.HIGHEST, evt -> {
       ServerPlayerEntity player = evt.getPlayer();
       Pokemon pokemon = evt.getPokemon();
       Boss boss = CobbleBosses.bossesConfig.getBoss(pokemon);
       if (boss != null) {
         pokemon.getPersistentData().remove(CobbleBosses.TAG_BOSS_ID);
-        pokemon.setLevel(boss.getDamageable().getLevel());
         boss.getRewards().giveRewards(player.getUuid());
       }
     });