+++ src/main/java/com/kingpixel/cobblebosses/model/Damageable.java
@@ -9,13 +9,13 @@ public class Damageable {
   private boolean enabled;
   private Double untilLifePercentage;
   private boolean catchable;
-  private int level;
+  private String properties;
 
   public Damageable() {
     this.enabled = false;
     this.untilLifePercentage = 10.0D;
     this.catchable = false;
-    this.level = 1;
+    this.properties = "level=10";
   }
 
   public boolean isDownLife(PokemonEntity pokemonEntity) {